requirements.txt
beautifulsoup4==4.11.1
Eel
googlesearch_python==1.3.0
hugchat==0.4.1
joblib==1.5.2
keras_facenet==0.3.2
langchain==1.1.0
langchain_community==0.4.1
langchain_core==1.1.0
langchain_google_genai==3.2.0
langgraph==1.0.4
markdown2==2.5.3
mtcnn
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
playsound==1.2.2
protobuf==6.33.1
psutil==5.9.1
pvporcupine==1.9.5
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycaw==20251023
python-dotenv==1.2.1
pyttsx3==2.90
pywhatkit==5.3
rapidfuzz==3.14.3
Requests==2.32.5
scikit_learn==1.7.2
screen_brightness_control==0.24.3
SpeechRecognition==3.8.1
tensorflow==2.20.0
typing_extensions==4.15.0
wikipedia==1.4.0
yt_dlp==2025.11.12
flask
opencv_python